delete impl files like reviewers said
