+++ app/src/androidTest/java/ch/epfl/cs311/wanderwave/viewmodel/TrackListViewModelTest.kt
@@ -1,5 +1,4 @@
 import ch.epfl.cs311.wanderwave.model.data.Track
-import ch.epfl.cs311.wanderwave.model.repository.TrackRepositoryImpl
 import ch.epfl.cs311.wanderwave.model.spotify.SpotifyController
 import ch.epfl.cs311.wanderwave.viewmodel.TrackListViewModel
 import io.mockk.every
